Extract run util
